Guest User can now access cart functionality.
- Guest users will now be able to also use the cart functionality,

- The use of session allows for guest users to also add to cart without it having to save anything to the database.

- Database will only be updated if user is registered and logged in as, the request.user.is_authenticated: will check if user is logged in or not and if not then it will request for the current session and save the updates to that.

- Could be a possibly good idea to include a way for users to be a guest add items to cart then log in and still have those added items to the cart when they were a guest be there.

+++ SunnybraeWeb/store/views.py
@@ -20,10 +20,10 @@ def product(request):
     page_number = request.GET.get('page')
     page_object = paginator.get_page(page_number)
     
-    print(page_object.number, " PAGE OBJECT PRINTED HERE ")
+    #print(page_object.number, " PAGE OBJECT PRINTED HERE ")
     
     context = {'page_object':page_object} #should render the page with x products as opposed to all products 
-    print ("CONTEXT ---------------- ",  context) 
+    #print ("CONTEXT ---------------- ",  context) 
     
     return render (request, 'store/product.html', context)
 
@@ -34,12 +34,31 @@ def faq(request):
 def cart(request):
     
     if request.user.is_authenticated:
-        customer = request.user.customer
+        customer, created = Customer.objects.get_or_create(user = request.user) #when a user registers have to make sure that a customer object is also created or retrieved so that that user can then shop 
         order, created = Order.objects.get_or_create(customer=customer, complete=False)
         items = order.orderitem_set.all()
     else:
+        session_cart = request.session.get('cart', {}) #if user is not authenticated i.e. a guest user then get the cart from the session
         items = []
         order = {"get_cart_total":0, "get_cart_items":0}
+        
+        
+        for productId, quantity in session_cart.items(): 
+            product = Product.objects.get(item_id=productId)
+            total = product.current_price * quantity
+            order['get_cart_items'] += quantity
+            order['get_cart_total'] += total 
+            item = {
+                'product':{
+                    'item_id': product.item_id,
+                    'item_name': product.item_name,
+                    'current_price': product.current_price
+                },
+                'quantity': quantity,
+                'get_total': total, 
+            }
+            items.append(item)
+        
         
     context = {'items':items, 'order': order}
     return render (request, 'store/cart.html', context)
@@ -50,26 +69,53 @@ def updateItem(request):
     productId = data['productId']
     action = data['action']
 
-    customer = request.user.customer
+    #customer = request.user.customer
     #print(customer)
     
-    product = Product.objects.get(item_id=productId) #match product with the product in the database 
-    order, created = Order.objects.get_or_create(customer=customer, complete=False) #get current order for customer and if order doesnt exist then create one 
+    if request.user.is_authenticated: #if the user is already registered 
+        customer, created = Customer.objects.get_or_create(user=request.user)
+        product = Product.objects.get(item_id=productId) #match product with the product in the database 
+        order, created = Order.objects.get_or_create(customer=customer, complete=False) #get current order for customer and if order doesnt exist then create one 
+        orderItem, created = OrderItem.objects.get_or_create(order=order, product=product) #get the OrderItem associated with the order if none exist then create one
+        
+        #actions will update the Database 
+        
+        if action == 'add': 
+            orderItem.quantity += 1
+        elif action == 'remove':
+            orderItem.quantity -= 1
+        elif action == 'delete':
+            orderItem.delete()
+            return JsonResponse('Item was deleted', safe=False)
 
-    orderItem, created = OrderItem.objects.get_or_create(order=order, product=product) #get the OrderItem associated with the order if none exist then create one
+        orderItem.save()
 
-    if action == 'add':
-        orderItem.quantity += 1
-    elif action == 'remove':
-        orderItem.quantity -= 1
-    elif action == 'delete':
-        orderItem.delete()
-        return JsonResponse('Item was deleted', safe=False)
-
-    orderItem.save()
-
-    if orderItem.quantity <= 0: #ensure that there is no items with a quanity value of zero in the cart 
-        orderItem.delete()
+        if orderItem.quantity <= 0: #ensure that there is no items with a quanity value of zero in the cart 
+            orderItem.delete()
+    
+    else: #if user is a guest 
+        session_cart = request.session.get('cart',{})
+        
+        #actions will only update the session which can be reset 
+        if action == 'add': 
+            if productId in session_cart:
+                session_cart[productId] += 1
+            else:
+                session_cart[productId] = 1
+                
+        elif action == 'remove':
+            if productId in session_cart:
+                session_cart[productId] -= 1
+                if session_cart[productId] <= 0:
+                    del session_cart[productId]
+        
+        elif action == 'delete':
+            if productId in session_cart:
+                del session_cart[productId]
+        
+        request.session['cart'] = session_cart
+        
+        
 
     return JsonResponse('Item was added', safe=False)
 
@@ -94,10 +140,11 @@ def user_register(request):
     if request.method == 'POST':
         form = RegisterForm(request.POST)
         if form.is_valid():
-            form.save()
+            user = form.save()
             Customer.objects.create(user=user, name=user.username, email=user.email)
             login (request, user)
             return redirect('product')
     else:
         form = RegisterForm()
     return render (request, 'store/register.html', {'form':form})
+
